refactor: move resource-server modules and add role facades
Move the remaining resource-server modules into
oauth2_provider/resource_server/ with sys.modules-alias shims at the old paths:

- resource_server/: backends, decorators, middleware

Wire up role-based facades so the public API can be imported by role:
- oauth2_provider.resource_server re-exports the RS decorators, backend,
  middleware, protected-resource views, and mixins.
- oauth2_provider.authorization_server re-exports the AS/token/introspect/
  device/DCR views and OAuth2Validator; provider modules (dcr, cimd, bcp,
  forms, admin) are importable directly.
- oauth2_provider.authorization_server.oidc re-exports the OP views.

View- and model-backed names load lazily (PEP 562 __getattr__) so importing a
role package never touches the app registry before django.setup(); verified by
importing all three role packages pre-setup without AppRegistryNotReady. Full
non-e2e suite passes (996).
