+++ oauth2_provider/backends.py
@@ -1,35 +1,21 @@
-from django.contrib.auth import get_user_model
-from django.core.exceptions import SuspiciousOperation
+"""Backward-compatible import shim.
 
-from .core.backends_oauthlib import get_oauthlib_core
+``oauth2_provider.backends`` moved to ``oauth2_provider.resource_server.backends`` when
+the package was reorganized by OAuth2 role. Importing from this old path still
+works but is deprecated and will be removed in django-oauth-toolkit 4.0.
+"""
 
+import sys
+import warnings
 
-UserModel = get_user_model()
-OAuthLibCore = get_oauthlib_core()
+from oauth2_provider.resource_server import backends as _moved
 
+warnings.warn(
+    "oauth2_provider.backends has moved to oauth2_provider.resource_server.backends. "
+    "The old import path is deprecated and will be removed in "
+    "django-oauth-toolkit 4.0.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-class OAuth2Backend:
-    """
-    Authenticate against an OAuth2 access token
-    """
-
-    def authenticate(self, request=None, **credentials):
-        if request is not None:
-            try:
-                valid, request = OAuthLibCore.verify_request(request, scopes=[])
-            except ValueError as error:
-                if str(error) == "Invalid hex encoding in query string.":
-                    raise SuspiciousOperation(error)
-                else:
-                    raise
-            else:
-                if valid:
-                    return request.user
-
-        return None
-
-    def get_user(self, user_id):
-        try:
-            return UserModel.objects.get(pk=user_id)
-        except UserModel.DoesNotExist:
-            return None
+sys.modules[__name__] = _moved
